auth service

+++ src/app/dashboard/components/toolbar/toolbar.component.ts
@@ -1,6 +1,7 @@
 import { Component, OnInit, Output, EventEmitter } from '@angular/core';
 import { JwtService } from '../../../core/services/jwt.service';
 import { Router } from '@angular/router';
+import { AuthService } from '../../../core/services/auth.service';
 
 @Component({
   selector: 'app-toolbar',
@@ -9,11 +10,21 @@ import { Router } from '@angular/router';
 })
 export class ToolbarComponent implements OnInit {
   @Output() toggleSidenav = new EventEmitter();
-  constructor(private jwtService: JwtService, private router: Router) {}
+  constructor(
+    private jwtService: JwtService,
+    private router: Router,
+    private authService: AuthService
+  ) {}
 
   ngOnInit() {}
   logOut() {
-    this.jwtService.destroyToken();
-    this.router.navigate(['login']);
+    const token = this.jwtService.getToken();
+    this.authService.logout().subscribe(response => {
+      if (response.success) {
+
+        this.jwtService.destroyToken();
+        this.router.navigate(['login']);
+      }
+    });
   }
 }
